TestApp/requirements.txt
python-dotenv==1.0.1
RPi.GPIO
pydantic==2.9.2
